requirements.txt
numpy==2.2.2
pandas==2.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
langchain-ollama==0.2.2
python-dotenv==1.0.0
streamlit==1.42.0
plotly==6.0.0